requirements.txt
<<<<<<< HEAD
Flask>=3.0.0
flask-cors>=4.0.0
Pillow>=10.0.0
rembg>=2.0.50
opencv-python-headless>=4.8.0
numpy>=2.3.0
waitress>=2.1.0
gunicorn>=21.0.0
onnxruntime>=1.16.0
=======
# Core dependencies (install first)
gunicorn==21.2.0
flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
waitress==3.0.0

# Background removal (primary method - AI-based, best quality)
# Note: rembg requires specific Python version compatibility
rembg==2.0.57

# OpenCV (fallback method if rembg fails)
opencv-python-headless==4.8.1.78
numpy==1.24.3
>>>>>>> d47371e (Fix Render deployment with proper gunicorn config)
